+++ Engine_Master/ModuleEditor.cpp
@@ -345,8 +345,9 @@ void ModuleEditor::drawScenePanel()
 {
 	ImGui::SetNextWindowSize(ImVec2(App->window->width * 0.6f, App->window->height * 0.625f));
 	ImGui::SetNextWindowPos(ImVec2(App->window->width * 0.2f, 85.0f));
-	if (ImGui::Begin(ICON_FA_FEATHER_ALT "Scene", NULL, ImGuiWindowFlags_NoCollapse | ImGuiWindowFlags_NoResize | ImGuiWindowFlags_NoMove))
+	if (ImGui::Begin(ICON_FA_FEATHER_ALT " Scene", NULL, ImGuiWindowFlags_NoCollapse | ImGuiWindowFlags_NoResize | ImGuiWindowFlags_NoMove))
 	{
+		updateWindowData("Scene");
 		App->renderer->drawSceneView();
 		ImGui::End();
 	}